Move the layout and graphics classes to double precision
git-svn-id: svn+ssh://svn.abisource.com/svnroot/abiword/branches/JMM-DOUBLEGRAPHICS-1@15385 bcba8976-2d24-0410-9c9c-aab3bd5fdfd6

+++ src/af/xap/unix/xap_UnixFrameImpl.cpp
@@ -381,8 +381,8 @@ static void s_gtkMenuPositionFunc(GtkMenu * /* menu */, gint * x, gint * y, gboo
 {
 	struct UT_Point * p = static_cast<struct UT_Point *>(user_data);
 
-	*x = p->x;
-	*y = p->y;
+	*x = (gint)p->x;
+	*y = (gint)p->y;
 	*push_in = TRUE ;
 }
